refactor: remove obsoletes repository api usages
