--- gravitee-apim-gateway/gravitee-apim-gateway-standalone/gravitee-apim-gateway-standalone-container/src/test/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-io.gravitee.repository.Repository=\
-io.gravitee.gateway.standalone.repository.MockRepository
