--- gravitee-apim-gateway/gravitee-apim-gateway-tests-sdk/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-io.gravitee.repository.Repository=\
-io.gravitee.gateway.MockRepository
