--- gravitee-apim-repository/gravitee-apim-repository-jdbc/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-io.gravitee.repository.Repository=io.gravitee.repository.jdbc.JdbcRepositoryProvider
